Create submodule

+++ backday-server/src/main/java/com/xebia/Action.java
@@ -1,8 +1,9 @@
 package com.xebia;
 
+import org.springframework.data.annotation.Id;
+
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-import org.springframework.data.annotation.Id;
 
 @XmlRootElement
 public class Action {
@@ -12,9 +13,9 @@ public class Action {
     private String id;
 
     @XmlElement
-    private Long value;
+    private Integer value;
 
-    public Action(final String id, final Long value) {
+    public Action(final String id, final Integer value) {
         this.id = id;
         this.value = value;
     }
@@ -31,7 +32,7 @@ public class Action {
         return id;
     }
 
-    public Long getValue() {
+    public Integer getValue() {
         return value;
     }
 }
